makefile: remove dependency from install target
There are no dependencies on the 'binary' target, so it will always
recompile. The README suggests running 'make' and then 'sudo make
install', but since the install target depends on the binary target,
this causes the build to be redone as root, which is both a waste of
compute time and a nuisance since it pollutes the source tree with files
owned by root.

A better fix for this would be to fill in the appropriate dependencies
for the binary target. This change is the quick+easy fix.