Enforce namespace deletion in internal KubeClient
This commit extends the KubeClient to ensure `shouldDeleteNamespace()`
waits for the complete deletion of a namespace before returning. This
guarantees that when the function is invoked, the namespace and its
resources are fully removed.

The change ensures reliable tests environments by preventing resource
collisions and enabling more e2e tests to be added.

Signed-off-by: Jean-Baptiste Louazel <jb@flex.ai>