feat: add log formatting (#435)
* feat: add log formatting

We introduce log formatting to support both text and json formats. This
can be configured through a CLI flag "log-format" or env var
"DCGM_EXPORTER_LOG_FORMAT".

Notably, the log format will capture any DCGM logging that is
configured.

We also move the automaxprocs call from an unused import to be
explicitly called when starting the dcgm exporter to ensure its logging
format honors the configuration.

* fix: use stderr output for slog handling

We use os.Stderr as using os.Stdout resulted in logs not being visible when
running the binary.

Tested via `go run ./cmd/dcgm-exporter` as well as with a Docker container
in a K8s cluster.

Change-Id: Icde060d9e5321a95d60f036ab619c6c7952f525e

* fix: fold in debug configuration to function

We fold in the debug logging setup to consolidate logging config into a single
function.

Additionally, we fix a formatting issues with the automaxprocs logging.

Change-Id: I958e50464cee7f1eca7799ad6ee2a6ba730f3234